refactor pt2
