+++ ui/litellm-dashboard/src/app/model_hub_table/page.tsx
@@ -7,6 +7,7 @@ export default function PublicModelHubTable() {
   const searchParams = useSearchParams()!;
   const key = searchParams.get("key");
   const [accessToken, setAccessToken] = useState<string | null>(null);
+  console.log("PublicModelHubTable accessToken:", accessToken);
 
   useEffect(() => {
     if (!key) {
