+++ ui/litellm-dashboard/src/components/AIHub/forms/make_agent_public_form.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { Modal, Form, Steps, Button, Checkbox } from "antd";
 import { Text, Title, Badge } from "@tremor/react";
-import { makeAgentsPublicCall } from "../networking";
-import NotificationsManager from "../molecules/notifications_manager";
+import { makeAgentsPublicCall } from "../../networking";
+import NotificationsManager from "../../molecules/notifications_manager";
 import { AgentHubData } from "@/components/AIHub/agent_hub_table_columns";
 
 const { Step } = Steps;
