+++ ui/litellm-dashboard/src/components/AIHub/forms/make_mcp_public_form.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import { Modal, Form, Steps, Button, Checkbox } from "antd";
 import { Text, Title, Badge } from "@tremor/react";
-import { makeMCPPublicCall } from "../networking";
-import NotificationsManager from "../molecules/notifications_manager";
-import { MCPServerData } from "@/components/AIHub/mcp_hub_table_columns";
+import { makeMCPPublicCall } from "../../networking";
+import NotificationsManager from "../../molecules/notifications_manager";
+import { MCPServerData } from "@/components/mcp_hub_table_columns";
 
 const { Step } = Steps;
