+++ ui/litellm-dashboard/src/components/AIHub/forms/make_model_public_form.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useCallback, useEffect } from "react";
 import { Modal, Form, Steps, Button, Checkbox } from "antd";
 import { Text, Title, Badge } from "@tremor/react";
-import { makeModelGroupPublic } from "../networking";
-import ModelFilters from "../model_filters";
-import NotificationsManager from "../molecules/notifications_manager";
+import { makeModelGroupPublic } from "../../networking";
+import ModelFilters from "../../model_filters";
+import NotificationsManager from "../../molecules/notifications_manager";
 
 const { Step } = Steps;
